refactor(util): rename hex-compare.ts → timing-safe-compare.ts + add constantTimeStringEqual
- Renames util/hex-compare.ts → util/timing-safe-compare.ts (the
  broader name reflects the module's expanded purpose). Export
  name sha256HexEqualConstantTime stays so existing callers update
  import paths only.
- Adds constantTimeStringEqual(a, b) export — Buffer-based UTF-8
  compare with the length-mismatch burn cycle from the
  to-be-deleted http-auth.ts helper. Canonical for invariant I10.
- Updates the two existing I10 callers (extensions/verify-
  extensions.ts and updater/updater.ts) to import from the new
  path. No logic changes.
- Removes the old util/hex-compare.test.ts; its coverage is folded
  into util/timing-safe-compare.test.ts (already committed in the
  preceding TDD-red commit).

The Task 2 util tests now pass (TDD green).
